Lazify loading of reactions to users mapping
Borrowed some ideas from code at https://medium.com/js-dojo/lazy-rendering-in-vue-to-improve-performance-dcccd445d5f

Generally the idea is that reaction avatars are now fetched only
when the reaction bar slides into view. This should lower the load
a bit on the server.

TODO: check there might be a glitch when adding a reaction